+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 * Added mmap.hpp and mmap.cpp
     * Added function to sort e820 map
 * Added file ‘init.hpp’ to klibc, with an inline function to call init functions for all other files in klibc which need to be initalized
-* Added `malloc`
+* Added `malloc`, `realloc`, and `free`.
 * Added linking to libgcc
 * Added `kentry()` and moved all critical initialization functions there, and renamed the original `kentry` `kmain`
 * Added algorithm.hpp and algorithm.cpp, for now only containing `foreach`